+++ saleor/graphql/payment/mutations/stored_payment_methods/__init__.py
@@ -1,0 +1,5 @@
+from .payment_method_request_delete import StoredPaymentMethodRequestDelete
+
+__all__ = [
+    "StoredPaymentMethodRequestDelete",
+]