+++ CHANGELOG.md
@@ -17,6 +17,13 @@ in 3.17. Use `PaymentSettings.defaultTransactionFlowStrategy` instead.
 
 ### Saleor Apps
 - Add `customerIpAddress` to `TRANSACTION_INITIALIZE_SESSION` and `TRANSACTION_PROCESS_SESSION` webhooks  #13718 by @korycins
+- Add `storedPaymentMethodRequestDelete` mutation - #13660 by @korycins
+- Add `PaymentSettings` to `Channel` - #13677 by @korycins
+
+### Saleor Apps
+- Add `STORED_PAYMENT_METHOD_DELETE_REQUESTED` webhook event - #13660 by @korycins
+- Add `NOTIFY_CUSTOMER` flag to `FulfillmentCreated` type - #13620, by @Air-t
+  - Inform apps if customer should be notified when fulfillment is created.
 
 ### Other changes
 - Fix error in variant available stock calculation - 13593 by @awaisdar001